backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
assemblyai
groq
elevenlabs
sqlalchemy
pydantic
python-multipart
requests
nltk
python-jose[cryptography]